migrate room to sqldelight and update hilt kapt to ksp

+++ app/src/main/java/com/example/sqldelightnoteapp/domain/add_edit_note_redux/AddEditNoteAction.kt
@@ -11,6 +11,6 @@ sealed class AddEditNoteAction: Action {
     data class ContentFocusChanged(val focusState: FocusState,val contentValue: String ): AddEditNoteAction()
     data class ColorChanged(val color: Int): AddEditNoteAction()
     object SaveNote: AddEditNoteAction()
-    data class FetchNoteById(val noteId: Int): AddEditNoteAction()
+    data class FetchNoteById(val noteId: Long): AddEditNoteAction()
     data class FetchNoteComplated(val note: Note): AddEditNoteAction()
 }
