+++ app/src/main/java/com/example/sqldelightnoteapp/domain/repository/NoteRepository.kt
@@ -7,7 +7,7 @@ interface NoteRepository {
 
     suspend fun getAllNotes(): List<Note>
 
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
 
     suspend fun insertNote(note: Note)
